+++ Amethyst/AST/Statements/ExecuteStatement.cs
@@ -5,61 +5,51 @@ using Geode.Types;
 
 namespace Amethyst.AST.Statements
 {
-	public class ExecuteStatement(
-		LocationRange loc,
-		IEnumerable<ExecuteStatementSubcommand> subCommands,
-		Statement stmt,
-		Statement? elseStmt) : Statement(loc)
-	{
-		public readonly Statement? Else = elseStmt;
-		public readonly Statement Statement = stmt;
-		public readonly ExecuteStatementSubcommand[] SubCommands = [.. subCommands];
+    public class ExecuteStatement(LocationRange loc, IEnumerable<ExecuteStatementSubcommand> subCommands, Statement stmt, Statement? elseStmt) : Statement(loc)
+    {
+        public readonly Statement? Else = elseStmt;
+        public readonly Statement Statement = stmt;
+        public readonly ExecuteStatementSubcommand[] SubCommands = [.. subCommands];
 
-		public override void Compile(FunctionContext ctx)
-		{
-			var chain = new ExecuteChain();
+        public override void Compile(FunctionContext ctx)
+        {
+            var chain = new ExecuteChain();
 
-			foreach (var i in SubCommands)
-			{
-				i.Compute(chain, ctx);
-			}
+            foreach (var i in SubCommands)
+            {
+                i.Compute(chain, ctx);
+            }
 
-			if (Else is not null)
-			{
-				ctx.Branch(chain, "execute", () => Statement.Compile(ctx), () => Else.Compile(ctx));
-			}
-			else
-			{
-				ctx.Branch(chain, "execute", () => Statement.Compile(ctx));
-			}
-		}
-	}
+            if (Else is not null)
+                ctx.Branch(chain, "execute", () => Statement.Compile(ctx), () => Else.Compile(ctx));
+            else
+                ctx.Branch(chain, "execute", () => Statement.Compile(ctx));
+        }
+    }
 
-	public abstract class ExecuteStatementSubcommand(LocationRange loc) : Node(loc)
-	{
-		public abstract void Compute(ExecuteChain chain, FunctionContext ctx);
-	}
+    public abstract class ExecuteStatementSubcommand(LocationRange loc) : Node(loc)
+    {
+        public abstract void Compute(ExecuteChain chain, FunctionContext ctx);
+    }
 
-	public class IfSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
-	{
-		public readonly Expression Expression = expr;
+    public class IfSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
+    {
+        public readonly Expression Expression = expr;
 
-		public override void Compute(ExecuteChain chain, FunctionContext ctx) => Expression.ExecuteChain(chain, ctx);
-	}
+        public override void Compute(ExecuteChain chain, FunctionContext ctx) => Expression.ExecuteChain(chain, ctx);
+    }
 
-	public class AsSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
-	{
-		public readonly Expression Expression = expr;
+    public class AsSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
+    {
+        public readonly Expression Expression = expr;
 
-		public override void Compute(ExecuteChain chain, FunctionContext ctx) =>
-			chain.Add(new AsChain(Expression.Execute(ctx, new TargetSelectorType())));
-	}
+        public override void Compute(ExecuteChain chain, FunctionContext ctx) => chain.Add(new AsChain(Expression.Execute(ctx, new TargetSelectorType())));
+    }
 
-	public class AtSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
-	{
-		public readonly Expression Expression = expr;
+    public class AtSubcommand(LocationRange loc, Expression expr) : ExecuteStatementSubcommand(loc)
+    {
+        public readonly Expression Expression = expr;
 
-		public override void Compute(ExecuteChain chain, FunctionContext ctx) =>
-			chain.Add(new AtChain(Expression.Execute(ctx, new TargetSelectorType())));
-	}
+        public override void Compute(ExecuteChain chain, FunctionContext ctx) => chain.Add(new AtChain(Expression.Execute(ctx, new TargetSelectorType())));
+    }
 }